update node info processor to include unschedulable nodes
This change updates the `Process` function of the node info processor
interface so that it can accept a second list of nodes. The second list
contains all the nodes that are not in the first list. This will allow
the mixed node info processor to properly detect unready and
unschedulable nodes for use as templates.